+++ src/contexts/index.js
@@ -1,1 +1,2 @@
 export * from "./overlay";
+export * from "./message-context";
